Split up masking.py
