+++ jax/interpreters/xla.py
@@ -42,6 +42,7 @@ from ..lib import xla_bridge as xb
 from ..lib import xla_client as xc
 from . import partial_eval as pe
 from . import ad
+from . import shapes
 from . import masking
 
 FLAGS = flags.FLAGS
@@ -947,7 +948,7 @@ device_put_p = core.Primitive('device_put')
 device_put_p.def_impl(_device_put_impl)
 pe.custom_partial_eval_rules[device_put_p] = lambda trace, x, **params: x
 ad.deflinear(device_put_p, lambda cotangent, **kwargs: [cotangent])
-masking.shape_rules[device_put_p] = lambda x, **_: x.shape
+shapes.shape_rules[device_put_p] = lambda x, **_: x.shape
 masking.defvectorized(device_put_p)
