+++ jax/lax/lax.py
@@ -27,7 +27,7 @@ from typing import Any
 import warnings
 
 import numpy as onp
-from jax.interpreters.masking import index_
+from jax.interpreters.shapes import to_index
 
 from ..util import partial, prod
 
@@ -42,12 +42,8 @@ from ..core import Primitive
 from ..abstract_arrays import (UnshapedArray, ShapedArray, ConcreteArray,
                                AbstractToken, array_types, make_shaped_array,
                                raise_to_shaped, abstract_token)
-from ..interpreters import partial_eval as pe
-from ..interpreters import xla
-from ..interpreters import pxla
-from ..interpreters import ad
-from ..interpreters import batching
-from ..interpreters import masking
+from ..interpreters import xla, pxla, ad, batching, shapes, masking, \
+  partial_eval as pe
 from ..util import curry, cache, safe_zip, unzip2, prod
 from ..tree_util import build_tree, tree_unflatten, tree_map
 from ..lib import pytree
@@ -78,7 +74,7 @@ def broadcast_shapes(*shapes):
 
 def _try_canonicalize_shape(shape):
   try:
-    return tuple(map(lambda x: masking.index_(x), shape))
+    return tuple(map(lambda x: shapes.to_index(x), shape))
   except (TypeError, AttributeError):
     return None
 
@@ -1087,7 +1083,7 @@ def iota(dtype, size):
   <https://www.tensorflow.org/xla/operation_semantics#iota>`_
   operator.
   """
-  size = index_(size)
+  size = to_index(size)
   dtype = dtypes.canonicalize_dtype(dtype)
   lazy_expr = lazy.iota(dtype, size)
   aval = ShapedArray((size,), dtype)
@@ -1528,7 +1524,7 @@ def standard_primitive(shape_rule, dtype_rule, name, translation_rule=None):
   prim.def_impl(partial(xla.apply_primitive, prim))
   prim.def_abstract_eval(partial(standard_abstract_eval, prim, shape_rule, dtype_rule))
   xla.translations[prim] = translation_rule or partial(standard_translate, name)
-  masking.shape_rules[prim] = shape_rule
+  shapes.shape_rules[prim] = shape_rule
   return prim
 
 
@@ -4116,7 +4112,7 @@ tie_in_p.def_abstract_eval(lambda x, y: raise_to_shaped(y))
 xla.translations[tie_in_p] = lambda c, x, y: y
 ad.deflinear(tie_in_p, _tie_in_transpose_rule)
 batching.primitive_batchers[tie_in_p] = _tie_in_batch_rule
-masking.shape_rules[tie_in_p] = lambda x, y: y.shape
+shapes.shape_rules[tie_in_p] = lambda x, y: y.shape
 masking.masking_rules[tie_in_p] = lambda vals, logical_shapes: vals[1]
